Changes for generate inputs
- Use usize over u64 as a small bitvector
- Add a prop test for truth table column generation

+++ src/main.rs
@@ -1,7 +1,7 @@
 use cartesian_genetic_programming::batch::BatchEval;
 
 fn main() {
-    let inputs : Vec<u64> = BatchEval::generate_inputs(3);
+    let inputs : Vec<usize> = BatchEval::generate_inputs(3);
     for input in inputs {
 	println!("{:#b}", input);
     }
